requirements.txt
aiogram==2.25.1
gspread
oauth2client
apscheduler
python-dotenv
